creating unit tests for new abinit reconstruction command
